requirement.txt
openai
gtts
moviepy
requests
pydub
pyaudio
google-generativeai
beautifulsoup4  # bs4 is typically installed as beautifulsoup4
python-ffmpeg
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client


